+++ backend/app/Http/Controllers/ServiceController.php
@@ -21,17 +21,31 @@ class ServiceController extends Controller
             'user_id' => auth()->user()->id,
         ]);
     }
-    public function update($id, Request $request) {
+    public function update($id, Request $request, Service $service) {
+        if (Gate::allows('authorization', $service)) {
+            try { //tikrina ar vartotojas yra prisijungęs, jeigu ne išveda klaidą
+                $user = auth()->userOrFail();
+            } catch (\Tymon\JWTAuth\Exceptions\UserNotDefinedException $e) {
+                return response()->json(['error' => 'Prašome prisijungti']);
+            }
+            Service::where('id', $id)->update($request->except(['_token']));
+        } else if (Gate::denies('authorization', $service)){
+            return response()->json(["error" => "Jūs neturite teisės"]);
+        }
+
+        return response()->json(["message" => "Paslauga sekmingai atnaujinta"]);
+    }
+    public function destroy(Request $request, Service $service) {
         try { //tikrina ar vartotojas yra prisijungęs, jeigu ne išveda klaidą
             $user = auth()->userOrFail();
         } catch(\Tymon\JWTAuth\Exceptions\UserNotDefinedException $e) {
             return response()->json(['error' => 'Prašome prisijungti']);
         }
-        Service::where('id',$id)->update($request->except(['_token']));
-        return response()->json(["message" => "Paslauga sekmingai atnaujinta"]);
-    }
-    public function destroy(Request $request, Service $service) {
-        $service->delete();
+        if (Gate::allows('authorization', $service)) {
+            $service->delete();
+        } else if (Gate::denies('authorization', $service)){
+            return response()->json(["error" => "Jūs neturite teisės"]);
+        }
         return response()->json(["message" => "Paslauga sekmingai ištrinta"]);
     }
 }
